+++ src/components/atomic-crm/dashboard/MobileDashboard.tsx
@@ -1,9 +1,8 @@
 import { useGetList, useTimeout } from "ra-core";
 import { Skeleton } from "@/components/ui/skeleton";
 
-import type { Contact, ContactNote } from "../types";
+import type { Account } from "../types";
 import { DashboardActivityLog } from "./DashboardActivityLog";
-import { DashboardStepper } from "./DashboardStepper";
 import { Welcome } from "./Welcome";
 import MobileHeader from "../layout/MobileHeader";
 import { MobileContent } from "../layout/MobileContent";
@@ -44,44 +43,18 @@ const Loading = () => (
 );
 
 export const MobileDashboard = () => {
-  const {
-    data: dataContact,
-    total: totalContact,
-    isPending: isPendingContact,
-  } = useGetList<Contact>("contacts", {
+  const { isPending } = useGetList<Account>("account_contacts", {
     pagination: { page: 1, perPage: 1 },
   });
-  const { total: totalContactNotes, isPending: isPendingContactNotes } =
-    useGetList<ContactNote>("contact_notes", {
-      pagination: { page: 1, perPage: 1 },
-    });
   const oneSecondHasPassed = useTimeout(1000);
-
-  const isPending = isPendingContact || isPendingContactNotes;
 
   if (isPending) {
     return oneSecondHasPassed ? <Loading /> : null;
   }
 
-  if (!totalContact) {
-    return (
-      <Wrapper>
-        <DashboardStepper step={1} />
-      </Wrapper>
-    );
-  }
-
-  if (!totalContactNotes) {
-    return (
-      <Wrapper>
-        <DashboardStepper step={2} contactId={dataContact?.[0]?.id} />
-      </Wrapper>
-    );
-  }
-
   return (
     <Wrapper>
-      <div className="grid grid-cols-1 md:grid-cols-12 gap-6 mt-1">
+      <div className="grid grid-cols-1 gap-6 mt-1">
         {import.meta.env.VITE_IS_DEMO === "true" ? <Welcome /> : null}
         <DashboardActivityLog />
       </div>